refactor(edda-conductor): split runner/sequential.rs into gate.rs + outcome.rs (GH-776)
Extract settle_gated_phase and the phase-result family so gate defect lanes
read a dedicated module instead of a 684-line run_plan. Verbatim move plus
handle_blocked / run_next_phase wrappers; tests stay on sequential::tests.

Base includes GH-751 (notify_gate_progress + restart-anchored deadline).
