+++ crates/edda-conductor/src/runner/mod.rs
@@ -1,5 +1,7 @@
 pub mod edda;
 pub mod event_log;
+pub mod gate;
 pub mod heartbeat;
 pub mod notify;
+pub mod outcome;
 pub mod sequential;
