+++ scripts/file-length-ceilings.txt
@@ -24,7 +24,8 @@ crates/edda-cli/src/main.rs 1456
 crates/edda-conductor/src/agent/codex_app_server.rs 1061
 crates/edda-conductor/src/agent/codex_rpc.rs 1378
 crates/edda-conductor/src/agent/pi_rpc.rs 1480
-crates/edda-conductor/src/runner/sequential.rs 6154
+crates/edda-conductor/src/runner/gate.rs 1020
+crates/edda-conductor/src/runner/sequential.rs 4593
 crates/edda-core/src/event.rs 2191
 crates/edda-derive/src/context/session.rs 1072
 crates/edda-ledger/src/ledger.rs 1927
